Made basic edit operations aware of characters above BMP and combining character sequences. (SF.net bug #3040720)
Specifically, the fixed operations are caret movements, selection,
delete, and backspace. Many other derived operations which use caret
positions or selection ranges also became safe unless the offsets were
altered by programs which are not aware of above mentioned characters.

Direct buffer operations are not in scope of this fix.


git-svn-id: http://jedit.svn.sourceforge.net/svnroot/jedit/jEdit/trunk@21259 6b1eeb88-9816-0410-afa2-b43733a0f04e